feat: add reusable password form
