+++ app/_locales/en_GB/messages.json
@@ -2004,6 +2004,9 @@
   "enterOptionalPassword": {
     "message": "Enter optional password"
   },
+  "enterPassword": {
+    "message": "Enter password"
+  },
   "enterPasswordContinue": {
     "message": "Enter password to continue"
   },
@@ -3365,7 +3368,7 @@
     "message": "New password"
   },
   "newPasswordPlaceholder": {
-    "message": "Use at least 8 characters"
+    "message": "Enter a strong password"
   },
   "newPrivacyPolicyActionButton": {
     "message": "Read more"
